refactor: move the total-metric card into the total-views and total-visitors widgets
