--- projects/packages/premium-analytics/packages/widgets-toolkit/src/components/stats-total-metric/index.ts
@@ -1,4 +1,0 @@
-export { StatsTotalMetricWidget } from './stats-total-metric';
-export type { StatsTotalMetricWidgetProps } from './stats-total-metric';
-export { useStatsTotalMetric } from './use-stats-total-metric';
-export type { StatsTotalMetricField, StatsTotalMetricState } from './use-stats-total-metric';
