+++ projects/packages/premium-analytics/packages/widgets-toolkit/src/helpers/index.ts
@@ -62,3 +62,4 @@ export { summaryCount } from './summary-count';
 export { toDay } from './to-day';
 export { defaultPeriodForInterval } from './default-period-for-interval';
 export { buildMetricTab, type MetricReport, type BuildMetricTabOptions } from './build-metric-tab';
+export { withoutComparison } from './without-comparison';
