+++ projects/packages/premium-analytics/widgets/total-visitors/package.json
@@ -4,10 +4,14 @@
 	"private": true,
 	"type": "module",
 	"dependencies": {
+		"@jetpack-premium-analytics/data": "link:../../packages/data",
+		"@jetpack-premium-analytics/externals": "link:../../packages/externals",
+		"@jetpack-premium-analytics/formatters": "link:../../packages/formatters",
 		"@jetpack-premium-analytics/widgets-toolkit": "link:../../packages/widgets-toolkit",
 		"@wordpress/i18n": "^6.9.0",
 		"@wordpress/icons": "^15.0.0",
-		"@wordpress/widget-primitives": "0.2.0"
+		"@wordpress/widget-primitives": "0.2.0",
+		"react": "18.3.1"
 	},
 	"devDependencies": {
 		"@wordpress/api-fetch": "7.51.0"
