+++ projects/packages/premium-analytics/widgets/total-visitors/style.module.css
@@ -1,0 +1,24 @@
+.root {
+	position: relative;
+	display: flex;
+	flex-direction: column;
+	block-size: 100%;
+	min-block-size: 0;
+}
+
+.body {
+	display: flex;
+	flex-direction: column;
+	block-size: 100%;
+	min-block-size: 0;
+	gap: var(--wpds-dimension-gap-sm);
+}
+
+/* `Sparkline` is the responsive variant, so it fills this box: let the box take
+   whatever the headline leaves, rather than pinning it to a fixed band that
+   strands empty space as the card grows. `min-block-size: 0` is what lets a
+   flex child shrink below its content size. */
+.chart {
+	flex: 1 1 auto;
+	min-block-size: 0;
+}
